+++ NEWS.md
@@ -1,3 +1,25 @@
+# optimizeR 0.2.0.9000
+
+* Inputs for `set_optimizer()` were renamed for clarity:
+  
+  * `opt_fun` -> `optimizer`
+  * `f` -> `objective`
+  * `p` -> `initial`
+  * `v` -> `value`
+  * `z` -> `parameter`
+  * `out_ign` -> `output_ignore`
+  * `test_par` -> `validation_settings`
+  
+* Elements for input `validation_settings` have been simplified.
+
+* Elements in the output of `apply_optimizer()` were renamed for clarity:
+
+  * `v` -> `value` 
+  * `z` -> `parameter`
+  * `time` -> `seconds` (now just a `numeric`) 
+  
+* Added pkgdown site: https://loelschlaeger.de/optimizeR
+
 # optimizeR 0.2.0
 
 * Some functions were renamed: